requirements.txt
# Core libraries
pandas>=2.1.0
numpy>=1.24.0

# Data collection
yfinance>=0.2.33
requests>=2.31.0  # For API requests

# Data visualization
matplotlib>=3.7.0
seaborn>=0.13.0
plotly>=5.18.0
bokeh>=3.1.0  # For interactive visualizations
mplfinance 

# Statistical analysis
scipy>=1.11.0
statsmodels>=0.14.0

# Machine learning (optional, for advanced analysis)
scikit-learn>=1.3.0

# Jupyter Notebook support
jupyter>=1.0.0
ipykernel>=6.0.0
nbconvert>=7.0.0  # For converting notebooks to other formats

# Additional utilities
tqdm>=4.65.0  # Progress bars
python-dotenv>=1.0.0  # Environment variables
openpyxl>=3.0.0  # For Excel file support
xlrd>=2.0.0  # For reading older Excel files

# Testing and debugging
pytest>=7.0.0  # For unit testing
pytest-cov>=4.0.0  # For test coverage

tensorflow>=2.12.0  # For deep learning
keras>=2.12.0  # High-level neural networks API
xgboost>=1.7.0  # For gradient boosting

streamlit>=1.29.0
ta>=0.10.0
newsapi-python>=0.2.7
textblob>=0.17.1